Address .NET 8 analyzer warnings
This change addresses the analyzer warnings introduced by .NET 8. One of the big changes here is the conversion from the `IFileShortcutHandler` interface to the `FileShortcutHandler` abstract class. The use of the interface methods was one of my first attempts at using that feature and I've decided that i'm not a huge fan of it in this particular case.