Upgrade to .NET 8 (#240)
- Upgrade to .NET 8 SDK and Runtime
- Address .NET 8 analyzer warnings
  - Addresses the analyzer warnings introduced by .NET 8. One of the big
changes here is the conversion from the `IFileShortcutHandler` interface
to the `FileShortcutHandler` abstract class called out for performance
implications. The use of the interface methods was one of my first
attempts at using that feature and I've decided that I'm not a huge fan
of it in this particular case.